Course List: Add new option to show current user's progress
On the Tutor LMS frontend user dashboard page (site.edu/dashboard),
the 1st dashboard entry and the entry "Enrolled courses" show the
list of enrolled courses.

This plugin's Course List element shows the same course cards.
The benefit of this element is of course that it is customizable.

Add an option to this plugin's Course List element to enable showing the
current user's course progress in the course list's course cards in the
same way the dashboard shows the user's course progress for each course.

The configuration setting is copied from the "Meta Data" toggle.
It defaults to be disabled to only show the user's course progress
when explicitly requested to not cause change to existing installations.

Additional minor fixes:
- Fix spelling of "Course Per Page" to "Courses Per Page"
- Fix Elementor Compatibility warning to show the version up to which
  this plugin is tested with, not "Unknown".