Fix crash of Elementor when $authordata is not set
Fix crash of Elementor when $authordata is not set, which
may be caused by slight inconsistencies of the database.

Tutor LMS handles these gracefully, so should do the Elementor
addon.

For example, when courses were imported but the author was not
imported beforehand, $authordata may not be set. Since it would
be False and not an array with an ID, this could crash the Elementor
editor when opening a page for editing.

Fix this crash by checking if $authordata is an array before
attempting to use $authordata->ID, and return an empty profile_url
in this case.